Extending zabbix_module_docker.so
I was working on CentOS 7, which changed the base of cgroups.
I was able to rebuild the zabbix_module_docker.so and test it. In my case it took everything in the folder "system.slice", where is also zabbix agent and some other stuff.
The docker container folders are here also. The Zabbix took all data from the discovery function and worked it through.
And those, which were real docker container are monitored well.
The filtering could be done directly in Zabbix, throgh regexp.

I would like to ask you to build the module and test it again against your testing enviroment.